implement high_activity finder
This code scans an image for edges denoted by high contrast to the surrounding
pixels. This produces a list of Rectangles, each containing an area with a
high probability of containing words. These rectangles will be individually
OCR'd, then considered in a location aware algorithm to map important info
such as name, title, company and contact info.

no footer